requirements.txt
fastapi
uvicorn
google-generativeai
pypdf
numpy
requests
python-dotenv